# chatty-backend/Docs/Swagger/post.swagger.yaml
openapi: 3.0.3
info:
  title: Chatty Post API
  description: API endpoints for posts (text, image, video)
  version: 1.0.0
servers:
  - url: /api/v1/posts
paths:
...
  /post/images/{page}:
    get:
      summary: Get posts with images (paginated)
      parameters:
        - in: path
          name: page
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: List of image posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
...
  /post/image/{postId}:
    put:
      summary: Update a post with image
      parameters:
        - in: path
          name: postId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateImagePostRequest'
      responses:
        '200':
          description: Image post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
        '404':
          description: Post not found
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UpdateImagePostRequest:
      type: object
      properties:
        content:
          type: string
        image:
          type: string
          format: binary
      required:
        - content
        - image
...